Make config files built in at build
We will later allow overrides, but this means that you
can go get wingo and run it without having to install a bunch
of files all over the place as root.

Create map of file paths to []byte
If we can't open
a file we can check in the FileMap to see if it's already there.
Use the FileMap for default scripts and data
If a file is not found, we can fall back to the built in file.
As long as we're at it
save them in there if we open them.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>